feat: add expenses metric line chart tooltip
